// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OrderType = {
  SALE: 'SALE',
  PURCHASE: 'PURCHASE'
} as const

export type OrderType = (typeof OrderType)[keyof typeof OrderType]


export const OrderStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const MovementType = {
  IN: 'IN',
  OUT: 'OUT',
  ADJUSTMENT: 'ADJUSTMENT'
} as const

export type MovementType = (typeof MovementType)[keyof typeof MovementType]


export const InvoiceStatus = {
  UNPAID: 'UNPAID',
  PAID: 'PAID',
  VOID: 'VOID'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]